fine-tuning/examples/PaddleHub/image_classification/requirements.txt
paddlepaddle-gpu
paddlehub
shapely
pyclipper
opencv-python
numpy
